added check before leave or kick

+++ src/main/java/net/cryptic_game/microservice/network/endpoint/NetworkMemberEndpoint.java
@@ -98,6 +98,10 @@ public class NetworkMemberEndpoint {
             return error("no_permissions");
         }
 
+        if (network.getOwner().equals(device)) {
+            return error("cannot_leave_own_network");
+        }
+
         for (Member member : Member.getMembers(network.getUUID())) {
             if (member.getDevice().equals(device)) {
                 member.delete();
